Items can now be used in code
Items can now be called for different functionalities and added a heal function to the baseunit class.

+++ Assets/Scripts/Items.cs
@@ -12,6 +12,6 @@ public class Items : ScriptableObject
     public Skills attachedSkill;
 
     public int itemUses;
-    public float itemHeal;
+    public int itemHeal;
 
 }
